% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{build_map_and_masterZ}
\alias{build_map_and_masterZ}
\title{Build MAP and masterZ matrices for DINE}
\usage{
build_map_and_masterZ(Z_in)
}
\arguments{
\item{Z_in}{List of N matrices, each with 2*K columns.}
}
\value{
A list containing the masterZ matrix and the MAP matrix, k (number of nodes) and t (number of timepoints).
}
\description{
Build MAP and masterZ matrices for DINE
}
