\name{DCENt-package}
\alias{DCENt-package}
\alias{DCENt}
\docType{package}
\title{
\packageTitle{DCENt}
}
\description{
\packageDescription{DCENt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DCENt}
\packageIndices{DCENt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DCENt}

Maintainer: \packageMaintainer{DCENt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
