% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Run DINE algorithm 1 or 2}
\usage{
estimate(
  X,
  y,
  Z,
  n0,
  k0,
  t0,
  algo = 2,
  max_itr = 200,
  convergence_cutoff = 5 * (10^(-5)),
  REML = FALSE,
  verbose = FALSE,
  timings = FALSE,
  n_fold = 5,
  n_threads = 1,
  threshold = NA
)
}
\arguments{
\item{X}{fixed effect matrix - should be in order all times for node 1 for subject 1, all times for node 2 for subject 1, and so on}

\item{y}{response vector - same order as X}

\item{Z}{list of matrices - one random effect matrix per subject. should have form 
[intercept time 1 0 ... ... 0] <- node 1
[intercept time 2 0 ... ... 0] <- node 1
[............................] <- node 1
[0 0 intercept time 1 0 ... 0] <- node 2}

\item{n0}{Integer. total number of subjects}

\item{k0}{Integer. total number of nodes}

\item{t0}{Integer. total number of timepoints}

\item{algo=2}{1 or 2. which algorithm to run}

\item{max_itr=200}{Integer. maximum number of iterations to run the algorithms}

\item{convergence_cutoff=5*10^-5.}{Double. relative difference in parameters used to determine algorithm convergence}

\item{REML=FALSE.}{Boolean. whether or not to use restricted ML for fitting in algorithm 2}

\item{verbose=FALSE.}{Boolean. For algo 2, prints current and previous D, E, and beta values at each iteration}

\item{timings=FALSE.}{Boolean. Whether or not to print the time each fitting takes for each iteration}

\item{n_fold=5}{Integer. Number of cross-validation fits to use to fit lambda for algorithm 2}

\item{n_threads=1}{Integer. Number of threads to use for loops. Set to 1 if you're running estimate in parallel already}

\item{threshold=NA}{Matrix of pre-computed thresholds to use for fitting D}
}
\value{
A named list containing: 
\itemize{
  \item \code{E}: The diagonal residual variance matrix.
  \item \code{D}: The thresholded random effects covariance matrix.
  \item \code{V}: The final composite covariance matrix.
  \item \code{beta}: The final fixed-effects coefficient vector.
  \item \code{timelength}: The amount of time the algorithm took to run
  \item \code{converged}: Logical indicating if the algorithm reached \code{convergence_cutoff} before \code{max_itr}.
  \item \code{sigma}: The estimated scaling variance parameter for algorithm 2. NA for algorithm 1
  \item \code{n_iter}: The total number of iterations run.
  \item \code{all_err}: Vector of convergence errors across all iterations.
  \item \code{MAP}: Integer matrix, n x kt describing which node/timepoint combinations each subject has
  \item \code{V_nonzeros_pct}: For algorithm 1, percentage of cells in the overall covariance V matrix that are zeros. NA for algorithm 2
  \item \code{threshold}: The final threshold matrix selected by cross-validation for algoritm 2. NA for algorithm 1.
}
}
\description{
Run DINE algorithm 1 or 2
}
