% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_masterZ}
\alias{estimate_masterZ}
\title{Run DINE algorithm 1 or 2}
\usage{
estimate_masterZ(
  X,
  y,
  masterZ,
  MAP,
  n0,
  k0,
  t0,
  algo = 2,
  max_itr = 200,
  convergence_cutoff = 5 * (10^(-5)),
  REML = FALSE,
  verbose = FALSE,
  timings = FALSE,
  n_fold = 5,
  n_threads = 1,
  threshold = NA
)
}
\arguments{
\item{X}{fixed effect matrix - should be in order all times for node 1 for subject 1, all times for node 2 for subject 1, and so on}

\item{y}{response vector - same order as X}

\item{masterZ}{overall random effect matrix for all timepoints across all subjects. Can be generated by build_map_and_masterZ}

\item{MAP}{Integer matrix, n x kt describing which node/timepoint combinations each subject has. Can be generated by build_map_and_masterZ}

\item{n0}{Integer. total number of subjects}

\item{k0}{Integer. total number of nodes}

\item{t0}{Integer. total number of timepoints}

\item{algo}{(optional) default value 2. 1 or 2. which algorithm to run}

\item{max_itr}{(optional) Integer. default value 100 Integer. maximum number of iterations to run the algorithms}

\item{convergence_cutoff}{(optional) default value 5*10^-5. Double. relative difference in parameters used to determine algorithm convergence}

\item{REML}{(optional) Boolean. default value FALSE. Boolean. whether or not to use restricted ML for fitting in algorithm 2}

\item{verbose}{(optional) Boolean. default value FALSE. Boolean. For algo 2, prints current and previous D, E, and beta values at each iteration}

\item{timings}{(optional) Boolean. default value FALSE. Boolean. Whether or not to print the time each fitting takes for each iteration}

\item{n_fold}{(optional) Integer. default value 5 Integer. Number of cross-validation fits to use to fit lambda for algorithm 2}

\item{n_threads}{(optional) Integer. default value 1 Integer. Number of threads to use for loops. Set to 1 if you're running estimate in parallel already}

\item{threshold}{(optional) Matrix. default value NA Matrix of pre-computed thresholds to use for fitting D}
}
\value{
A named list containing: 
\itemize{
  \item \code{E}: The diagonal residual variance matrix.
  \item \code{D}: The thresholded random effects covariance matrix.
  \item \code{V}: The final composite covariance matrix.
  \item \code{beta}: The final fixed-effects coefficient vector.
  \item \code{timelength}: The amount of time the algorithm took to run
  \item \code{converged}: Logical indicating if the algorithm reached \code{convergence_cutoff} before \code{max_itr}.
  \item \code{sigma}: The estimated scaling variance parameter for algorithm 2. NA for algorithm 1
  \item \code{n_iter}: The total number of iterations run.
  \item \code{all_err}: Vector of convergence errors across all iterations.
  \item \code{V_nonzeros_pct}: For algorithm 1, percentage of cells in the overall covariance V matrix that are zeros. NA for algorithm 2
  \item \code{threshold}: The final threshold matrix selected by cross-validation for algoritm 2. NA for algorithm 1.
}
}
\description{
Run DINE algorithm 1 or 2
}
