\name{PLoCONE-package}
\alias{PLoCONE-package}
\alias{PLoCONE}
\docType{package}
\title{
\packageTitle{PLoCONE}
}
\description{
\packageDescription{PLoCONE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PLoCONE}
\packageIndices{PLoCONE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PLoCONE}

Maintainer: \packageMaintainer{PLoCONE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
