\name{DINE-package}
\alias{DINE-package}
\alias{DINE}
\docType{package}
\title{
\packageTitle{DINE}
}
\description{
\packageDescription{DINE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DINE}
\packageIndices{DINE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DINE}

Maintainer: \packageMaintainer{DINE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
